\name{demineur}
\alias{demineur}
\title{Application de démineur}
\description{
  Application Shiny de démineur. Cette application permet aux utilisateurs de jouer au jeu de démineur avec différentes configurations et niveaux de difficulté.
}
\usage{
demineur
}
\format{
  Un objet shinyApp représentant l'application de démineur.
}
\examples{
# Ceci est un exemple d'utilisation dans le cadre d'une application Shiny
# library(shiny)
# library(shinyjs)
# library(dbplyr)
#
# source("ui.R")
# source("server.R")
# source("fonction.R")
#
# demineur = shinyApp(ui = ui, server = server)
# runApp(demineur)
}
