\name{server}
\alias{server}
\title{Fonction serveur pour l'application de démineur}
\description{
  La fonction serveur pour l'application Shiny de démineur. Cette fonction gère la logique du jeu, y compris la génération de la grille, le calcul du temps restant, le décompte des drapeaux et la détection des victoires et des défaites.
}
\usage{
server(input, output, session)
}
\arguments{
  \item{input}{Liste d'entrées de l'application Shiny.}
  \item{output}{Liste de sorties de l'application Shiny.}
  \item{session}{Objet de session de l'application Shiny.}
}
\value{
  Aucune valeur de retour. La fonction modifie directement les entrées, sorties et l'état de la session de l'application Shiny.
}
\examples{
# Ceci est un exemple d'utilisation dans le cadre d'une application Shiny
# library(shiny)
# library(shinyjs)
# library(dbplyr)
#
# source("ui.R")
# source("fonction.R")
#
# shinyApp(ui = ui, server = server)
}
