\name{ui}
\alias{ui}
\title{Interface utilisateur pour l'application de démineur}
\description{
  Interface utilisateur pour l'application Shiny de démineur. Cette interface comprend la configuration du jeu, les options d'action, un bouton pour réinitialiser le jeu, les règles du jeu, un chronomètre et un décompte des drapeaux placés.
}
\usage{
ui
}
\format{
  Un objet fluidPage contenant les éléments de l'interface utilisateur.
}
\examples{
# Ceci est un exemple d'utilisation dans le cadre d'une application Shiny
# library(shiny)
# library(shinyjs)
# library(dbplyr)
#
# source("ui.R")
# source("fonction.R")
#
# shinyApp(ui = ui, server = server)
}
